Split catalog suite
